qcacmn: Change the data type of average/roundup tx rate
The datatype of average and round up tx rate was uint32_t which truncates
the upper byte of the average and round up tx rate value.

changing the datatype of average and round up tx rate from uint32_t to
uint64_t.

CRs-Fixed: 2409751

+++ dp/cmn_dp_api/dp_ratetable.h
@@ -65,14 +65,14 @@ enum CMN_BW_TYPES {
 #define DUMMY_MARKER	  0
 #define DP_ATH_RATE_IN(c)  (DP_ATH_EP_MUL((c), DP_ATH_RATE_EP_MULTIPLIER))
 
-static inline int dp_ath_rate_lpf(int _d, int _e)
+static inline int dp_ath_rate_lpf(uint64_t _d, int _e)
 {
 	_e = DP_ATH_RATE_IN((_e));
 	return (((_d) != DUMMY_MARKER) ? ((((_d) << 3) + (_e) - (_d)) >> 3) :
 			(_e));
 }
 
-static inline int dp_ath_rate_out(int _i)
+static inline int dp_ath_rate_out(uint64_t _i)
 {
 	int _mul = DP_ATH_RATE_EP_MULTIPLIER;
